+++ core/validatorapi/validatorapi.go
@@ -24,7 +24,6 @@ import (
 	"github.com/attestantio/go-eth2-client/spec"
 	eth2p0 "github.com/attestantio/go-eth2-client/spec/phase0"
 	"github.com/coinbase/kryptology/pkg/signatures/bls/bls_sig"
-	ssz "github.com/ferranbt/fastssz"
 	"go.opentelemetry.io/otel/trace"
 
 	"github.com/obolnetwork/charon/app/errors"
@@ -526,27 +525,4 @@ func (c Component) getProposerPubkey(ctx context.Context, slot eth2p0.Slot) (cor
 	}
 
 	return pubkey, nil
-}
-
-// MerkleEpoch wraps epoch to implement ssz.HashRoot.
-type MerkleEpoch eth2p0.Epoch
-
-func (m MerkleEpoch) HashTreeRoot() ([32]byte, error) {
-	b, err := ssz.HashWithDefaultHasher(m)
-	if err != nil {
-		return [32]byte{}, errors.Wrap(err, "hash default epoch")
-	}
-
-	return b, nil
-}
-
-func (m MerkleEpoch) HashTreeRootWith(hh *ssz.Hasher) error {
-	indx := hh.Index()
-
-	// Field (1) 'Epoch'
-	hh.PutUint64(uint64(m))
-
-	hh.Merkleize(indx)
-
-	return nil
 }
