+++ testutil/validatormock/validatormock.go
@@ -37,10 +37,10 @@ type Eth2Provider interface {
 	eth2client.AttestationDataProvider
 	eth2client.AttestationsSubmitter
 	eth2client.AttesterDutiesProvider
-	eth2client.DomainProvider
-	eth2client.ProposerDutiesProvider
 	eth2client.BeaconBlockProposalProvider
 	eth2client.BeaconBlockSubmitter
+	eth2client.DomainProvider
+	eth2client.ProposerDutiesProvider
 	eth2client.SlotsPerEpochProvider
 	eth2client.SpecProvider
 	eth2client.ValidatorsProvider
